pyproject.toml
[project]
name = "sistema-avaliacoes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beanie>=2.0.1",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "fastapi-pagination>=0.15.6",
    "motor>=3.7.1",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.40.0",
]
